test: payment 약속확정 & 수수료 결제 기능 테스트 작성 Resolves:#12

+++ src/main/java/com/example/demo/payment/dto/RequestPayDto.java
@@ -11,10 +11,10 @@ public class RequestPayDto {
     private String orderUid;
     private String itemName;
     private String buyerName;
-    private Long paymentPrice;
+    private int paymentPrice;
 
     @Builder
-    public RequestPayDto(String orderUid, String itemName, String buyerName, Long paymentPrice) {
+    public RequestPayDto(String orderUid, String itemName, String buyerName, int paymentPrice) {
         this.orderUid = orderUid;
         this.itemName = itemName;
         this.buyerName = buyerName;
